Relocate Craft.ComponentSelectInput to a yii2-adapter compat asset
Core no longer instantiates Craft.ComponentSelectInput — every core CP
surface renders the self-booting <craft-component-select> web component.
The legacy jQuery class survived only for the componentSelect.twig
`jsClass` escape hatch (plugin subclasses).

Move the real 681-line implementation verbatim out of the craftcms-legacy
CP bundle into legacy/web/assets/cpcompat/component-select-input.js so the
bundle can drop it while plugin subclasses keep a working base class. It
defines its class at top-level eval (not on DOMContentLoaded like
cp-compat.js) so a plugin's Craft.Foo = Craft.ComponentSelectInput.extend()
can run whenever its own bundle loads, and is registered ahead of
cp-compat.js via CpCompatAsset. Drop the Craft.js import and fix the
component-select README/index comments that claimed the class still ships
in the legacy bundle.
