+++ packages/craftcms-legacy/cp/src/Craft.js
@@ -42,7 +42,6 @@ import './js/CP.js';
 import './js/CardsElementIndexView.js';
 import './js/CategoryIndex.js';
 import './js/Chart.js';
-import './js/ComponentSelectInput.js';
 import './js/ContentBlockEditor.js';
 import './js/CustomizeSourcesModal.js';
 import './js/DataTableSorter.js';
